Move Dataframe classes to its own folder
